httpHandlerをmainから分離しmainをcollectorから一般的なものに
